Refactored legacy code for plotting station-array
